requirements.txt
pymongo
websockets~=11.0.3
tf2-sku-to-name